Refactored player data serializer to not delete data on load but rather when the player quits without saving. Added return function if player does not have a valid player data to save

+++ Assets/Scripts/PlayerData/PlayerDataSerializer.cs
@@ -11,14 +11,15 @@ public static class PlayerDataSerializer
     public static void Save()
     {
         if (PlayerData.MatchData == null)
+        {
             Debug.LogError("Unable to save data, no player match data set");
+            return;
+        }
         
         var data = PlayerData.MatchData;
-
         var jsonData = JsonUtility.ToJson(data);
 
         Directory.CreateDirectory(savePath);
-        
         File.WriteAllText(savePath + saveFileName, jsonData, Encoding.Default);
     }
 
@@ -28,8 +29,12 @@ public static class PlayerDataSerializer
         
         var jsonData = File.ReadAllText(savePath + saveFileName, Encoding.Default);
         PlayerData.MatchData = JsonUtility.FromJson<PlayerMatchData>(jsonData);
-        
-        File.Delete(savePath + saveFileName);
+    }
+
+    public static void DeleteSaveData()
+    {
+        if (PlayerHasData())
+            File.Delete(savePath + saveFileName);
     }
 
     public static bool PlayerHasData() => File.Exists(savePath + saveFileName);
